Factor out schema module
