+++ src/odf/metadata/src/schema/arrow_encoding.rs
@@ -7,8 +7,8 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use super::dtos_generated::*;
-use crate::dtos::data_schema_impl::UnsupportedSchema;
+use super::schema_impl::UnsupportedSchema;
+use crate::dtos::*;
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
